Rename workflow->executor
